实现视频播放

+++ app/src/main/cpp/native-lib.cpp
@@ -25,8 +25,6 @@ extern "C"
 JNIEXPORT void JNICALL
 Java_com_bochao_ffmpeg_SaoZhuPlayer_native_1start(JNIEnv *env, jobject thiz, jstring path_,
                                                   jobject surface) {
-
-
     // 记录结果
     int result;
     // R1 Java String -> C String
@@ -131,11 +129,9 @@ Java_com_bochao_ffmpeg_SaoZhuPlayer_native_1start(JNIEnv *env, jobject thiz, jst
                 return;
             }
             // 数据格式转换
-            result = sws_scale(
-                    data_convert_context,
-                    (const uint8_t *const *) frame->data, frame->linesize,
-                    0, videoHeight,
-                    rgba_frame->data, rgba_frame->linesize);
+            result = sws_scale(data_convert_context, (const uint8_t *const *) frame->data,
+                               frame->linesize, 0, videoHeight, rgba_frame->data,
+                               rgba_frame->linesize);
             if (result < 0) {
                 ALOGE("Player Error : data convert fail");
                 return;
@@ -152,7 +148,7 @@ Java_com_bochao_ffmpeg_SaoZhuPlayer_native_1start(JNIEnv *env, jobject thiz, jst
                 for (int h = 0; h < videoHeight; h++) {
                     memcpy(bits + h * window_buffer.stride * 4,
                            out_buffer + h * rgba_frame->linesize[0],
-                           rgba_frame->linesize[0]);
+                           static_cast<size_t>(rgba_frame->linesize[0]));
                 }
                 ANativeWindow_unlockAndPost(native_window);
             }
@@ -178,126 +174,4 @@ Java_com_bochao_ffmpeg_SaoZhuPlayer_native_1start(JNIEnv *env, jobject thiz, jst
     avformat_close_input(&format_context);
     // 释放 R1
     env->ReleaseStringUTFChars(path_, path);
-
-//*************************************************************************************************************
-
-//    ANativeWindow *nativeWindow = ANativeWindow_fromSurface(env, surface);
-//    const char *path = env->GetStringUTFChars(path_, 0);
-//    // FFmpeg 视频绘制
-//    //初始化网络
-//    avformat_network_init();
-//    //总上下文
-//    AVFormatContext *formatContext = avformat_alloc_context();
-//    AVDictionary *opts = NULL;
-//    av_dict_set(&opts, "timeout", "3000000", 0);
-//    int ret = avformat_open_input(&formatContext, path, NULL, &opts);
-//    ALOGE("------%i", ret);
-//    if (ret) {
-//        ALOGE("Player Error : Can not open video file");
-//        return;
-//    }
-//    //视频流
-//    int video_stream_idx = -1;
-//    ret = avformat_find_stream_info(formatContext, NULL);
-//    ALOGE("------%i", ret);
-//    if (ret != 0) {
-//        ALOGE("get video info fail");
-//        return;
-//    }
-//    for (int i = 0; i < formatContext->nb_streams; ++i) {
-//        if (formatContext->streams[i]->codecpar->codec_type == AVMEDIA_TYPE_VIDEO) {
-//            video_stream_idx = i;
-//            break;
-//        }
-//    }
-//    if (video_stream_idx == -1) {
-//        ALOGE("Player Error : Can not find video stream");
-//        return;
-//    }
-//    ALOGE("1");
-//    ALOGE("------%i", video_stream_idx);
-//    //视频流索引
-//    AVCodecParameters *codecpar = formatContext->streams[video_stream_idx]->codecpar;
-//    ALOGE("2");
-//    //解码器
-//    AVCodec *dec = avcodec_find_decoder(codecpar->codec_id);
-//    if (dec == NULL) {
-//        ALOGE("无法解码");
-//        return;
-//    }
-//    ALOGE("3");
-//    //获取解码器的上下文对象
-//    AVCodecContext *codecContext = avcodec_alloc_context3(NULL);
-//    ALOGE("4");
-//
-//    //解码  获取yuv数据
-//    ret = avcodec_open2(codecContext, dec, NULL);
-//    if (ret < 0) {
-//        ALOGE("解码器无法打开");
-//        return;
-//    }
-//    ALOGE("5");
-//    AVPacket *packet = av_packet_alloc();
-//    ALOGE("6");
-//    //从视频流中读取数据包
-//    // 获取视频的宽高
-//    int videoWidth = codecContext->width;
-//    int videoHeight = codecContext->height;
-//    struct SwsContext *swsContext = sws_getContext(
-//            videoWidth, videoHeight, codecContext->pix_fmt,
-//            videoWidth, videoHeight, AV_PIX_FMT_RGBA,
-//            SWS_BICUBIC, NULL, NULL, NULL);
-//    ALOGE("7");
-//
-//    ANativeWindow_setBuffersGeometry(nativeWindow, codecContext->width, codecContext->height,
-//                                     WINDOW_FORMAT_RGBA_8888);
-//    ALOGE("8");
-//
-//    ANativeWindow_Buffer outBuffer;
-//    while (av_read_frame(formatContext, packet) >= 0) {
-//        avcodec_send_packet(codecContext, packet);
-//        ALOGE("8-1");
-//        AVFrame *frame = av_frame_alloc();
-//        ALOGE("8-2");
-//        ret = avcodec_receive_frame(codecContext, frame);
-//        ALOGE("%i", ret);
-//        ALOGE("8-3");
-//        if (ret == AVERROR(EAGAIN)) {
-//            break;
-//        } else if (ret < 0) {
-//            break;
-//        }
-//        ALOGE("8-4");
-//        //接收的容器
-//        uint8_t *dst_data[4];
-//        //每一行的首地址
-//        int dst_linesize[4];
-//        av_image_alloc(dst_data, dst_linesize, codecContext->width, codecContext->height,
-//                       AV_PIX_FMT_RGBA, 1);
-//        //绘制
-//        sws_scale(swsContext, frame->data, frame->linesize, 0, frame->height, dst_data,
-//                  dst_linesize);
-//
-//        //加锁
-//        ANativeWindow_lock(nativeWindow, &outBuffer, NULL);
-//
-//        //渲染
-//        uint8_t *firstWindow = static_cast<uint8_t *>(outBuffer.bits);
-//        //输入源
-//        uint8_t *src_data = dst_data[0];
-//        //拿到一行的多少个字节RGBA
-//        int destStride = outBuffer.stride * 4;
-//        int src_linesize = dst_linesize[0];
-//        for (int i = 0; i < outBuffer.height; ++i) {
-//            //内存拷贝进行渲染
-//            memccpy(firstWindow + i * destStride, src_data + i * src_linesize, destStride, NULL);
-//        }
-//        //解锁
-//        ANativeWindow_unlockAndPost(nativeWindow);
-//        usleep(1000 * 16);
-//        av_frame_free(&frame);
-//
-//    }
-//    ALOGE("9");
-//    env->ReleaseStringUTFChars(path_, path);
 }
